Fix for ShowerEventSimdataTree xmax_pos_shc setter - type from float32 to float64

+++ grand/io/root_trees.py
@@ -5132,7 +5132,7 @@ class ShowerEventSimdataTree(MotherEventTree):
 
     @xmax_pos_shc.setter
     def xmax_pos_shc(self, value):
-        self._xmax_pos_shc = np.array(value).astype(np.float32)
+        self._xmax_pos_shc = np.array(value).astype(np.float64)
         self._tree.SetBranchAddress("xmax_pos_shc", self._xmax_pos_shc)
 
     @property
